+++ irc3/compat.py
@@ -8,3 +8,10 @@ if PY3:  # pragma: no cover
 else:  # pragma: no cover
     text_type = unicode
     string_types = basestring
+
+try:  # pragma: no cover
+    import asyncio
+    from asyncio.queues import Queue
+except ImportError:  # pragma: no cover
+    import trollius as asyncio  # NOQA
+    from trollius.queues import Queue  # NOQA